+++ lib/Models/WebMapServiceCatalogItem.ts
@@ -60,6 +60,7 @@ import WebMapServiceCapabilities, {
   CapabilitiesLayer,
   getRectangleFromLayer
 } from "./WebMapServiceCapabilities";
+import WebMapServiceCatalogGroup from "./WebMapServiceCatalogGroup";
 
 const dateFormat = require("dateformat");
 
@@ -160,8 +161,14 @@ class GetCapabilitiesStratum extends LoadableStratum(
               );
 
         if (layerStyle !== undefined && layerStyle.legend !== undefined) {
+          let url = this.catalogItem.supportsColorScaleRange
+            ? `${layerStyle.legend.url}&colorscalerange=${this.catalogItem.colorScaleRange}`
+            : layerStyle.legend.url;
           result.push(
-            <StratumFromTraits<LegendTraits>>(<unknown>layerStyle.legend)
+            createStratumInstance(LegendTraits, {
+              url,
+              urlMimeType: layerStyle.legend.urlMimeType
+            })
           );
         }
 
@@ -516,6 +523,38 @@ class GetCapabilitiesStratum extends LoadableStratum(
     }
   }
 
+  // TODO - There is possibly a better way to do this
+  @computed
+  get isThredds(): boolean {
+    if (
+      this.catalogItem.url &&
+      (this.catalogItem.url.indexOf("thredds") > -1 ||
+        this.catalogItem.url.indexOf("tds") > -1)
+    ) {
+      return true;
+    }
+    return false;
+  }
+
+  // TODO - Geoserver also support NCWMS via a plugin, just need to work out how to detect that
+  @computed
+  get isNcWMS(): boolean {
+    if (this.catalogItem.isThredds) return true;
+    return false;
+  }
+
+  @computed
+  get isEsri(): boolean {
+    if (this.catalogItem.url !== undefined)
+      return this.catalogItem.url.indexOf("MapServer/WMSServer") > -1;
+    return false;
+  }
+
+  @computed
+  get supportsColorScaleRange(): boolean {
+    return this.catalogItem.isNcWMS;
+  }
+
   @computed
   get discreteTimes(): { time: string; tag: string | undefined }[] | undefined {
     const result = [];
@@ -651,6 +690,8 @@ class WebMapServiceCatalogItem
     i18next.t("models.webMapServiceCatalogItem.getCapabilitiesUrl")
   ];
 
+  _webMapServiceCatalogGroup: undefined | WebMapServiceCatalogGroup = undefined;
+
   static defaultParameters = {
     transparent: true,
     format: "image/png",
@@ -670,6 +711,14 @@ class WebMapServiceCatalogItem
   // TODO
   get isMappable() {
     return true;
+  }
+
+  @computed
+  get colorScaleRange(): string | undefined {
+    if (this.supportsColorScaleRange) {
+      return `${this.colorScaleMinimum},${this.colorScaleMaximum}`;
+    }
+    return undefined;
   }
 
   async createGetCapabilitiesStratumFromParent(
@@ -942,10 +991,13 @@ class WebMapServiceCatalogItem
         ...dimensionParameters
       };
 
+      if (this.supportsColorScaleRange) {
+        parameters.COLORSCALERANGE = this.colorScaleRange;
+      }
+
       if (isDefined(this.styles)) {
         parameters.styles = this.styles;
       }
-
       Object.assign(parameters, diffModeParameters);
 
       const maximumLevel = scaleDenominatorToLevel(this.minScaleDenominator);
@@ -958,7 +1010,10 @@ class WebMapServiceCatalogItem
         "width",
         "height",
         "bbox",
-        "layers"
+        "layers",
+        // This is here as a temporary fix until Cesium implements this fix
+        // https://github.com/CesiumGS/cesium/issues/9021
+        "version"
       ];
 
       const baseUrl = queryParametersToRemove.reduce(