+++ CHANGES.md
@@ -8,6 +8,8 @@ Change Log
 * Update magda error message
 * Add a short report section if trying to view a `3d-tiles` item in a 2d map.
 * Fix bug in `Terria.interpretStartData`.
+* Add `ThreddsCatalogGroup` model.
+* Port `supportsColorScaleRange`, `colorScaleMinimum` and `colorScaleMaximimum` from `master` to `WebMapServiceCatalogItem` model.
 * Ported MapboxVectorTileCatalogItem ("mvt").
 * [The next improvement]